[FIX] Login: Replace old login with the new OAuth Login
