+++ src/js/services/blobIDS.service.js
@@ -66,6 +66,7 @@ module.factory('rpBlobIDS', ['$rootScope', '$http', '$q', function($scope, $http
 
       }, function (response) {
         if (response.status === 401) {
+          // TODO check where should this be placed
           // backend token is broken
           store.remove('backend_token');
         }
