+++ src/js/services/id.service.js
@@ -511,7 +511,7 @@ module.factory(
     $scope.$on('$routeChangeStart', function(ev, next) {
       if (!$scope.loginStatus) {
         var tab = next.tabName;
-        var allTabs = ['login', 'loginoauth', 'register', 'migrate', 'recover', '404', 'privacypolicy', 'tou'];
+        var allTabs = ['login', 'register', 'migrate', 'recover', '404', 'privacypolicy', 'tou'];
         if (allTabs.indexOf(tab) !== -1) {
           $scope.showLogin = false;
           $scope.showRegister = false;
