+++ src/templates/tabs/loginOld.jade
@@ -1,0 +1,101 @@
+section.col-xs-12.content(ng-controller="LoginCtrl")
+  .row.auth-form-container(ng-hide="$routeParams.to")
+    .col-xs-12.col-sm-4.col-md-8.hidden-xs
+      .welcomeText(l10n) Welcome to Ripple Trade
+      hr
+      p.heroText(l10n) Trade hundreds of assets, from XRP to crypto to gold to fiat.
+    .auth-form-wrapper.col-xs-12.col-sm-6.col-md-4
+      .tabAdjust
+        .signUpWrapper
+          h2.signUpTop
+            a(href='#register', l10n-inc) Sign Up
+        .loginWrapper
+          h2.loginTop(l10n) Log In
+      form(name='loginForm', ng-submit='submitForm()')
+        .auth-attention.text-center.spaceAdjust(ng-show="verifyStatus")
+          .status(ng-show="'verifying' === verifyStatus", l10n) Verifying...
+          .status(ng-show="'verified' === verifyStatus", l10n) You have successfully verified your email address.
+          .status(ng-show="'error' === verifyStatus", l10n) Email verification token is invalid. It has either expired or has been resent. Please check your inbox for the most recent verification email.
+        include ../tabs/login/form
+        .text-status(ng-show='status')
+          span.loggingIn(ng-bind="status")
+          br
+          .backend.error(ng-repeat="message in backendMessages")
+            span.error(ng-bind="message")
+        .submit-btn-container
+          button.btn.btn-submit.btn-block.btn-success#loginBtn(type='submit',
+            ng-disabled="ajax_loading || loginForm.$invalid", ng-hide="twoFactor")
+            img(src="img/button-s.png", class="loader", ng-show="ajax_loading")
+            span(class="loading_text", ng-show="ajax_loading", l10n) Logging in
+            span(ng-hide="ajax_loading", l10n) Log In
+          button.btn.btn-submit.btn-block.btn-success(type='submit',
+            ng-disabled="ajax_loading || !token", ng-show="twoFactor")
+            img(src="img/button-s.png", class="loader", ng-show="ajax_loading")
+            span(class="loading_text", l10n) Verify
+          a.btn.btn-block.btn-cancel(class="2fa-cancel", ng-click="cancel2FA()", ng-show="twoFactor", l10n) cancel
+      .row.switch-mode-wrapper
+        .switch-mode-link-container.bottomText(l10n) Already have a ripple.com/client account?&#32;
+          a(href='#migrate', l10n-inc) Migrate
+      .submit-btn-container-recovery
+        a.recover-btn(href="#recover/{{username}}", ng-show="showRecover")
+          button.btn.btn-block.btn-primary(ng-click="")
+            span(l10n) Account Recovery
+      button.btn.btn-link.recover(ng-click="goto_recover()", ng-hide="showRecover", l10n) Account Recovery
+
+
+  .row.action-login(ng-if="$routeParams.to")
+    .col-xs-12.col-sm-6.col-md-6
+      div.info(ng-show="'send' == $route.current.tabName")
+        p.literal(ng-bind="$routeParams.label")
+        div(ng-hide="$routeParams.amount", l10n)
+          p(l10n-inc) You're sending money to
+          .address {{ $routeParams.to | rpripplename:{tilde:true} }}
+          | .
+        div(ng-show="$routeParams.amount", l10n)
+          p(l10n-inc) You're sending
+          .amount(ng-show="$routeParams.amount")
+            span.number {{$routeParams.amount | rpamount:{xrp_human: true} }}
+            span.currency {{$routeParams.amount | rpcurrency}}
+          p(l10n-inc) &#32;to&#32;
+          .address {{ $routeParams.to | rpripplename:{tilde:true} }}
+          | .
+      div.info(ng-show="'trust' == $route.current.tabName")
+        p.literal {{$routeParams.label}}
+        div(l10n)
+          p(l10n-inc) You're connecting to the gateway
+          .address {{ $routeParams.to | rpripplename:{tilde:true} }}
+          div(ng-show="$routeParams.currency", l10n-inc)
+            | for
+            .amount
+              span.currency(ng-bind="$routeParams.currency")
+          div(ng-show="$routeParams.amount", l10n-inc)
+            | with a limit of
+            .amount
+              span.currency {{$routeParams.amount | rpamount}} {{$routeParams.amount | rpcurrency}}
+          | .
+
+      div.info(ng-show="'contacts' == $route.current.tabName")
+        p.literal(ng-bind="$routeParams.label")
+        span(l10n)
+          p(l10n-inc) You're adding
+          .address {{ $routeParams.to | rpripplename:{tilde:true} }}
+          p(l10n-inc) to your contacts list
+
+    .col-xs-12.col-sm-6.col-md-5.col-lg-4.col-lg-offset-1.auth-form-container
+      .auth-form-wrapper
+        form(name='loginForm', ng-submit='submitForm()')
+          include ../tabs/login/form
+          .row
+            .col-xs-12.text-status(ng-show='status')
+              span(ng-bind="status")
+              br
+              .backend(ng-repeat="message in backendMessages")
+                span(ng-bind="message")
+          .row
+            .col-xs-12
+              button.btn.btn-lg.btn-submit.btn-block.btn-primary(type='submit', ng-disabled="loginForm.$invalid || ajax_loading", l10n)
+                img(src="img/button-s.png", class="loader", ng-show="ajax_loading")
+                |  Log In
+          p.literal.hint(l10n) You can confirm this transaction on the next page.
+          p.literal(l10n) New to {{productName}}?&#32;
+            a(href='#/register', l10n-inc) Create a wallet
