--- src/templates/tabs/loginoauth.jade
@@ -1,5 +1,0 @@
-section.col-xs-12.content(ng-controller="LoginOAuthCtrl")
-  .row.auth-form-container()
-    .col-xs-12.col-sm-4.col-md-8
-      h1
-        a(href="{{globalOptions.backend_url}}/auth/login") Connect
